+++ test/python/ibmq/test_ibmq_job_states.py
@@ -11,14 +11,17 @@
 """IBMQJob states test-suite."""
 
 import unittest
-import time
 from contextlib import suppress
 from qiskit.providers.jobstatus import JobStatus
-from qiskit.providers.ibmq.ibmqjob import IBMQJobPreQobj, IBMQJob, API_FINAL_STATES
-from qiskit.providers.ibmq.api import ApiError
+from qiskit.providers.ibmq.ibmqjob import IBMQJobPreQobj, API_FINAL_STATES
 from qiskit.providers import JobError, JobTimeoutError
+from qiskit.test.mock import (FakeRueschlikon, BaseFakeAPI, UnknownStatusAPI, ValidatingAPI,
+                              ErrorWhileValidatingAPI, NonQueuedAPI, ErrorWhileCreatingAPI,
+                              ErrorWhileRunningAPI, QueuedAPI, RejectingJobAPI, UnavailableRunAPI,
+                              ThrowingAPI, ThrowingNonJobRelatedErrorAPI, ThrowingGetJobAPI,
+                              CancellableAPI, NonCancellableAPI, ErroredCancellationAPI,
+                              new_fake_qobj, _auto_progress_api)
 from .jobtestcase import JobTestCase
-from .._mockutils import new_fake_qobj, FakeBackend
 
 
 class TestIBMQJobStates(JobTestCase):
@@ -248,309 +251,15 @@ class TestIBMQJobStates(JobTestCase):
                     else:
                         self.assertFalse(self._current_api.get_job.called)
 
-    # TODO: Once qobj results come by default from all the simulator
-    # backends, move to integration tests in test_result.py
-    def test_qobj_result(self):
-        job = self.run_with_api(QObjResultAPI(), job_class=IBMQJob)
-
-        self.wait_for_initialization(job)
-        self._current_api.progress()
-        result = job.result()
-        self.assertEqual(result.success, True)
-        self.assertEqual(result.get_counts('Bell state'),
-                         {'00': 480, '11': 490, '01': 20, '10': 34})
-        self.assertEqual(result.get_counts('Bell state XY'),
-                         {'00': 29, '11': 15, '01': 510, '10': 480})
-        self.assertEqual(len(result.results), 2)
-
     def run_with_api(self, api, job_class=IBMQJobPreQobj):
         """Creates a new ``IBMQJobPreQobj`` instance running with the provided API
         object.
         """
-        backend = FakeBackend()
+        backend = FakeRueschlikon()
         self._current_api = api
         self._current_qjob = job_class(backend, None, api, False, qobj=new_fake_qobj())
         self._current_qjob.submit()
         return self._current_qjob
-
-
-def _auto_progress_api(api, interval=0.2):
-    """Progress a `BaseFakeAPI` instacn every `interval` seconds until reaching
-    the final state.
-    """
-    with suppress(BaseFakeAPI.NoMoreStatesError):
-        while True:
-            time.sleep(interval)
-            api.progress()
-
-
-class BaseFakeAPI():
-    """Base class for faking the IBM-Q API."""
-
-    class NoMoreStatesError(Exception):
-        """Raised when it is not possible to progress more."""
-
-    _job_status = []
-
-    _can_cancel = False
-
-    def __init__(self):
-        self._state = 0
-        self.config = {'hub': None, 'group': None, 'project': None}
-        if self._can_cancel:
-            self.config.update({
-                'hub': 'test-hub',
-                'group': 'test-group',
-                'project': 'test-project'
-            })
-
-    def get_job(self, job_id):
-        if not job_id:
-            return {'status': 'Error', 'error': 'Job ID not specified'}
-        return self._job_status[self._state]
-
-    def get_status_job(self, job_id):
-        summary_fields = ['status', 'error', 'infoQueue']
-        complete_response = self.get_job(job_id)
-        return {key: value for key, value in complete_response.items()
-                if key in summary_fields}
-
-    def run_job(self, *_args, **_kwargs):
-        time.sleep(0.2)
-        return {'id': 'TEST_ID'}
-
-    def cancel_job(self, job_id, *_args, **_kwargs):
-        if not job_id:
-            return {'status': 'Error', 'error': 'Job ID not specified'}
-        return {} if self._can_cancel else {
-            'error': 'testing fake API can not cancel'}
-
-    def progress(self):
-        if self._state == len(self._job_status) - 1:
-            raise self.NoMoreStatesError()
-        self._state += 1
-
-
-class UnknownStatusAPI(BaseFakeAPI):
-    """Class for emulating an API with unknown status codes."""
-
-    _job_status = [
-        {'status': 'UNKNOWN'}
-    ]
-
-
-class ValidatingAPI(BaseFakeAPI):
-    """Class for emulating an API with job validation."""
-
-    _job_status = [
-        {'status': 'VALIDATING'},
-        {'status': 'RUNNING'}
-    ]
-
-
-class ErrorWhileValidatingAPI(BaseFakeAPI):
-    """Class for emulating an API processing an invalid job."""
-
-    _job_status = [
-        {'status': 'VALIDATING'},
-        {'status': 'ERROR_VALIDATING_JOB'}
-    ]
-
-
-class NonQueuedAPI(BaseFakeAPI):
-    """Class for emulating a successfully-completed non-queued API."""
-
-    _job_status = [
-        {'status': 'RUNNING'},
-        {'status': 'COMPLETED', 'qasms': []}
-    ]
-
-
-class ErrorWhileCreatingAPI(BaseFakeAPI):
-    """Class emulating an API processing a job that errors while creating
-    the job.
-    """
-
-    _job_status = [
-        {'status': 'ERROR_CREATING_JOB'}
-    ]
-
-
-class ErrorWhileRunningAPI(BaseFakeAPI):
-    """Class emulating an API processing a job that errors while running."""
-
-    _job_status = [
-        {'status': 'RUNNING'},
-        {'status': 'ERROR_RUNNING_JOB', 'error': 'Error running job'}
-    ]
-
-
-class QueuedAPI(BaseFakeAPI):
-    """Class for emulating a successfully-completed queued API."""
-
-    _job_status = [
-        {'status': 'RUNNING', 'infoQueue': {'status': 'PENDING_IN_QUEUE'}},
-        {'status': 'RUNNING'},
-        {'status': 'COMPLETED'}
-    ]
-
-
-class RejectingJobAPI(BaseFakeAPI):
-    """Class for emulating an API unable of initializing."""
-
-    def run_job(self, *_args, **_kwargs):
-        return {'error': 'invalid qobj'}
-
-
-class UnavailableRunAPI(BaseFakeAPI):
-    """Class for emulating an API throwing before even initializing."""
-
-    def run_job(self, *_args, **_kwargs):
-        time.sleep(0.2)
-        raise ApiError()
-
-
-class ThrowingAPI(BaseFakeAPI):
-    """Class for emulating an API throwing in the middle of execution."""
-
-    _job_status = [
-        {'status': 'RUNNING'}
-    ]
-
-    def get_job(self, job_id):
-        raise ApiError()
-
-
-class ThrowingNonJobRelatedErrorAPI(BaseFakeAPI):
-    """Class for emulating an scenario where the job is done but the API
-    fails some times for non job-related errors.
-    """
-
-    _job_status = [
-        {'status': 'COMPLETED'}
-    ]
-
-    def __init__(self, errors_before_success=2):
-        super().__init__()
-        self._number_of_exceptions_to_throw = errors_before_success
-
-    def get_job(self, job_id):
-        if self._number_of_exceptions_to_throw != 0:
-            self._number_of_exceptions_to_throw -= 1
-            raise ApiError()
-
-        return super().get_job(job_id)
-
-
-class ThrowingGetJobAPI(BaseFakeAPI):
-    """Class for emulating an API throwing in the middle of execution. But not in
-       get_status_job() , just in get_job().
-       """
-
-    _job_status = [
-        {'status': 'COMPLETED'}
-    ]
-
-    def get_status_job(self, job_id):
-        return self._job_status[self._state]
-
-    def get_job(self, job_id):
-        raise ApiError('Unexpected error')
-
-
-class CancellableAPI(BaseFakeAPI):
-    """Class for emulating an API with cancellation."""
-
-    _job_status = [
-        {'status': 'RUNNING'},
-        {'status': 'CANCELLED'}
-    ]
-
-    _can_cancel = True
-
-
-class NonCancellableAPI(BaseFakeAPI):
-    """Class for emulating an API without cancellation running a long job."""
-
-    _job_status = [
-        {'status': 'RUNNING'},
-        {'status': 'RUNNING'},
-        {'status': 'RUNNING'}
-    ]
-
-
-class ErroredCancellationAPI(BaseFakeAPI):
-    """Class for emulating an API with cancellation but throwing while
-    trying.
-    """
-
-    _job_status = [
-        {'status': 'RUNNING'},
-        {'status': 'RUNNING'},
-        {'status': 'RUNNING'}
-    ]
-
-    _can_cancel = True
-
-    def cancel_job(self, job_id, *_args, **_kwargs):
-        return {'status': 'Error', 'error': 'test-error-while-cancelling'}
-
-
-# TODO: Remove once qobj results come by default from all the simulator
-# backends.
-class QObjResultAPI(BaseFakeAPI):
-    """Class for emulating a successfully-completed non-queued API."""
-
-    _job_status = [
-        {'status': 'RUNNING'},
-        {
-            'status': 'COMPLETED',
-            'qObjectResult': {
-                'backend_name': 'ibmqx2',
-                'backend_version': '1.1.1',
-                'job_id': 'XC1323XG2',
-                'qobj_id': 'Experiment1',
-                'success': True,
-                'status': 'COMPLETED',
-                'results': [
-                    {
-                        'header': {
-                            'name': 'Bell state',
-                            'memory_slots': 2,
-                            'creg_sizes': [['c', 2]],
-                            'clbit_labels': [['c', 0], ['c', 1]],
-                            'qubit_labels': [['q', 0], ['q', 1]]
-                        },
-                        'shots': 1024,
-                        'status': 'DONE',
-                        'success': True,
-                        'data': {
-                            'counts': {
-                                '0x0': 480, '0x3': 490, '0x1': 20, '0x2': 34
-                            }
-                        }
-                    },
-                    {
-                        'header': {
-                            'name': 'Bell state XY',
-                            'memory_slots': 2,
-                            'creg_sizes': [['c', 2]],
-                            'clbit_labels': [['c', 0], ['c', 1]],
-                            'qubit_labels': [['q', 0], ['q', 1]]
-                        },
-                        'shots': 1024,
-                        'status': 'DONE',
-                        'success': True,
-                        'data': {
-                            'counts': {
-                                '0x0': 29, '0x3': 15, '0x1': 510, '0x2': 480
-                            }
-                        }
-                    }
-                ]
-            }
-        }
-    ]
 
 
 if __name__ == '__main__':